Изменение типа продаж с int на long

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
-        int[] sales = new int[]{500, 1000, 5000, 50};
+        long[] sales = new long[]{500, 1000, 5000, 50};
         SalesManager salesManager = new SalesManager(sales);
-        int max = salesManager.max();
+        long max = salesManager.max();
         System.out.println(max);
     }
 }
